Create ZooGame.java
ZooGame is an interactive game that demonstrates the ability for a program to learn. It is a very basic implementation of Artificial Intelligence using trees as a data structure, and recursion to traverse the trees. Requires ZooGameNode.class in the same folder to run from command line, so you have to complile both Zoogame.java, and ZooGameNode.java. Have fun!